third_party/WebKit/LayoutTests/fast/dom/Range/missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
var range = document.createRange();
var node = document.body;
[
    "compareBoundaryPoints()",
    "compareBoundaryPoints(0)",
    "comparePoint()",
    "comparePoint(node)",
    "createContextualFragment()",
    "insertNode()",
    "intersectsNode()",
    "isPointInRange()",
    "isPointInRange(node)",
    "selectNode()",
    "selectNodeContents()",
    "setEnd()",
    "setEnd(node)",
    "setEndAfter()",
    "setEndBefore()",
    "setStart()",
    "setStart(node)",
    "setStartAfter()",
    "setStartBefore()",
    "surroundContents()"
].forEach(function(expr)
{
    shouldThrow("range." + expr);
});
</script>
</body>
</html>
