third_party/blink/web_tests/fast/dom/Range/missing-arguments-expected.txt

PASS range.compareBoundaryPoints() threw exception TypeError: Failed to execute 'compareBoundaryPoints' on 'Range': 2 arguments required, but only 0 present..
PASS range.compareBoundaryPoints(0) threw exception TypeError: Failed to execute 'compareBoundaryPoints' on 'Range': 2 arguments required, but only 1 present..
PASS range.comparePoint() threw exception TypeError: Failed to execute 'comparePoint' on 'Range': 2 arguments required, but only 0 present..
PASS range.comparePoint(node) threw exception TypeError: Failed to execute 'comparePoint' on 'Range': 2 arguments required, but only 1 present..
PASS range.createContextualFragment() threw exception TypeError: Failed to execute 'createContextualFragment' on 'Range': 1 argument required, but only 0 present..
PASS range.insertNode() threw exception TypeError: Failed to execute 'insertNode' on 'Range': 1 argument required, but only 0 present..
PASS range.intersectsNode() threw exception TypeError: Failed to execute 'intersectsNode' on 'Range': 1 argument required, but only 0 present..
PASS range.isPointInRange() threw exception TypeError: Failed to execute 'isPointInRange' on 'Range': 2 arguments required, but only 0 present..
PASS range.isPointInRange(node) threw exception TypeError: Failed to execute 'isPointInRange' on 'Range': 2 arguments required, but only 1 present..
PASS range.selectNode() threw exception TypeError: Failed to execute 'selectNode' on 'Range': 1 argument required, but only 0 present..
PASS range.selectNodeContents() threw exception TypeError: Failed to execute 'selectNodeContents' on 'Range': 1 argument required, but only 0 present..
PASS range.setEnd() threw exception TypeError: Failed to execute 'setEnd' on 'Range': 2 arguments required, but only 0 present..
PASS range.setEnd(node) threw exception TypeError: Failed to execute 'setEnd' on 'Range': 2 arguments required, but only 1 present..
PASS range.setEndAfter() threw exception TypeError: Failed to execute 'setEndAfter' on 'Range': 1 argument required, but only 0 present..
PASS range.setEndBefore() threw exception TypeError: Failed to execute 'setEndBefore' on 'Range': 1 argument required, but only 0 present..
PASS range.setStart() threw exception TypeError: Failed to execute 'setStart' on 'Range': 2 arguments required, but only 0 present..
PASS range.setStart(node) threw exception TypeError: Failed to execute 'setStart' on 'Range': 2 arguments required, but only 1 present..
PASS range.setStartAfter() threw exception TypeError: Failed to execute 'setStartAfter' on 'Range': 1 argument required, but only 0 present..
PASS range.setStartBefore() threw exception TypeError: Failed to execute 'setStartBefore' on 'Range': 1 argument required, but only 0 present..
PASS range.surroundContents() threw exception TypeError: Failed to execute 'surroundContents' on 'Range': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
